+++ src/subcommands/runtime/helpers/AliasGenerator.ts
@@ -183,65 +183,3 @@ export class AliasGenerator {
       .filter((alias): alias is BuiltAlias => alias !== null);
   }
 }
-
-/**
- * Llama.cpp specific generator that requires GPU framework to be included.
- */
-export class LlamaCppAliasGenerator extends AliasGenerator {
-  /**
-   * Returns base component sets for Llama.cpp engines
-   * @returns Array of field sets
-   */
-  protected override getBaseAliasComponentSets(): Set<AliasField>[] {
-    // Force llama.cpp engines to include the GPU framework for improved comprehension
-    return [
-      // No engine-only alias
-      new Set<AliasField>(["engine", "gpuFramework"]),
-      new Set<AliasField>(["engine", "gpuFramework", "platform"]),
-      new Set<AliasField>(["engine", "gpuFramework", "platform", "cpuArchitecture"]),
-      new Set<AliasField>([
-        "engine",
-        "gpuFramework",
-        "platform",
-        "cpuArchitecture",
-        "cpuInstructionSetExtensions",
-      ]),
-    ];
-  }
-}
-
-/**
- * MLX specific generator that maps mlx-llm to mlx-engine.
- */
-export class MlxAliasGenerator extends AliasGenerator {
-  /**
-   * Maps MLX engine names for display purposes.
-   * @param engineName - The original engine name
-   * @returns The mapped engine name
-   */
-  protected override mapEngineName(engineName: string): string {
-    // Rename mlx-llm -> mlx-engine for improved comprehension
-    return engineName === "mlx-llm" ? "mlx-engine" : engineName;
-  }
-}
-
-/**
- * Factory for creating appropriate alias generators based on engine type
- */
-export class AliasGeneratorFactory {
-  /**
-   * Creates the appropriate alias generator for the specified engine type.
-   * @param engineType - The engine type to create a generator for
-   * @returns Specialized alias generator for the engine type
-   */
-  static createGenerator(engineType: string): AliasGenerator {
-    switch (engineType) {
-      case "llama.cpp":
-        return new LlamaCppAliasGenerator();
-      case "mlx-llm":
-        return new MlxAliasGenerator();
-      default:
-        return new AliasGenerator();
-    }
-  }
-}
